// FRONTEND/project_uteq/src/app/models/proyecto.ts
export interface Proyecto {
  nombre: string;
  descripcion: string;
  tipo_articulo: string; // Asegúrate de usar el mismo nombre que en el backend
  usuario_id: any; // Asegúrate de usar el mismo nombre que en el backend
  plantilla_id: number; // Asegúrate de usar el mismo nombre que en el backend
}